requirements.txt
pycryptodome
UnityPy
msgpack
pillow